requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=2.7.0
ollama
python-dotenv
openpyxl